+++ sdk/typespec/typespec_client_core/src/lib.rs
@@ -15,6 +15,7 @@ pub mod fs;
 pub mod http;
 #[cfg(feature = "json")]
 pub mod json;
+pub mod serde;
 pub mod sleep;
 pub mod stream;
 #[cfg(feature = "xml")]