+++ sdk/typespec/typespec_client_core/src/async_runtime/mod.rs
@@ -106,7 +106,10 @@ pub trait AsyncRuntime: Send + Sync {
     ///
     fn spawn(&self, f: TaskFuture) -> SpawnedTask;
 
-    fn sleep(&self, duration: std::time::Duration) -> TaskFuture;
+    fn sleep(
+        &self,
+        duration: std::time::Duration,
+    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'static>>;
 }
 
 static ASYNC_RUNTIME_IMPLEMENTATION: OnceLock<Arc<dyn AsyncRuntime>> = OnceLock::new();