+++ sdk/core/azure_core/CHANGELOG.md
@@ -8,7 +8,11 @@
 
 ### Breaking Changes
 
+- Removed `AccessToken::is_expired()`
+
 ### Bugs Fixed
+
+- `BearerTokenCredentialPolicy` returns an error when a proactive token refresh attempt fails
 
 ### Other Changes