+++ sdk/core/azure_core/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features Added
 
 - Added `#[safe]` attribute helper for `SafeDebug` derive macro to show or hide types and members as appropriate.
+- Added module `serde` from `typespec_client_core`.
 
 ### Breaking Changes