+++ sdk/core/azure_core/CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Features Added
 
 - Added `#[safe]` attribute helper for `SafeDebug` derive macro to show or hide types and members as appropriate.
+- Added `get_async_runtime()` and `set_async_runtime()` to allow customers to replace
+the asynchronous runtime used by the Azure SDK.
 
 ### Breaking Changes