+++ sdk/core/azure_core/CHANGELOG.md
@@ -5,10 +5,16 @@
 ### Features Added
 
 - Added `#[safe]` attribute helper for `SafeDebug` derive macro to show or hide types and members as appropriate.
+- Added `Page` trait to facilitate the `ItemIterator`.
+- Added `PageIterator` to asynchronously iterator all pages.
 
 ### Breaking Changes
 
-- Removed `AccessToken::is_expired()`
+- A `Pager` now asynchronously iterates over items across all pages. Call `Pager::into_pages()` to get a `PageIterator` to asynchronously iterate over all pages.
+- Removed `AccessToken::is_expired()`.
+- Renamed `PagerResult::Continue` to `More` and its `continuation` field to `next`.
+- Renamed `PagerResult::Complete` to `Done`.
+- Renamed `PageStream` to `ItemIterator`.
 
 ### Bugs Fixed